Fixes #6514: Move to_fts_query to tribler_common.utilities from tribler_gui.utilities
