+++ src/tribler-common/tribler_common/utilities.py
@@ -1,4 +1,5 @@
 import os
+import re
 import sys
 from pathlib import Path
 from urllib.parse import urlparse
@@ -21,3 +22,13 @@ def uri_to_path(uri):
     parsed = urlparse(uri)
     host = "{0}{0}{mnt}{0}".format(os.path.sep, mnt=parsed.netloc)
     return Path(host) / url2pathname(parsed.path)
+
+
+fts_query_re = re.compile(r'\w+', re.UNICODE)
+
+
+def to_fts_query(text):
+    if not text:
+        return ""
+    words = fts_query_re.findall(text)
+    return ' '.join(f'"{word}"' for word in words) + '*'
