+++ src/tribler-gui/tribler_gui/widgets/searchresultswidget.py
@@ -7,11 +7,12 @@ from PyQt5 import uic
 from PyQt5.QtCore import pyqtSignal
 
 from tribler_common.sentry_reporter.sentry_mixin import AddBreadcrumbOnShowMixin
+from tribler_common.utilities import to_fts_query
 
 from tribler_core.components.metadata_store.db.serialization import CHANNEL_TORRENT, COLLECTION_NODE, REGULAR_TORRENT
 
 from tribler_gui.tribler_request_manager import TriblerNetworkRequest
-from tribler_gui.utilities import connect, get_ui_file_path, to_fts_query, tr
+from tribler_gui.utilities import connect, get_ui_file_path, tr
 from tribler_gui.widgets.tablecontentmodel import SearchResultsModel
 
 widget_form, widget_class = uic.loadUiType(get_ui_file_path('search_results.ui'))
